requirements.txt
alembic==1.0.8
blinker==1.4
Click==7.0
cymysql==0.9.13
Flask==1.0.2
Flask-Caching==1.7.0
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
WTForms==2.2.1
